return code

+++ packages/plugin-posclient-api/src/graphql/schema/orders.ts
@@ -35,7 +35,7 @@ export const types = `
   type PosOrderItem {
     ${commonFields}
     productId: String!
-    count: Float!
+    count: Int!
     orderId: String!
     unitPrice: Float
     discountAmount: Float
@@ -117,7 +117,7 @@ export const types = `
   input OrderItemInput {
     _id: String
     productId: String!
-    count: Float!
+    count: Int!
     unitPrice: Float!
     isPackage: Boolean
     isTake: Boolean
